Throw an error if we're attempting to update properties for some reason
I've noticed link rewriting will occasionally but reproducibly replace
a node's actual ID with the backlink replacement text when processing a
file.

This changeset tests whether `looking-at` actually spotted a link at
point. This _might_ address the issue, but I'm not sure exactly what's
causing it. For now, throw at error if we're in a property drawer.